[iOS] Shift playlist item processing/delegates into tab helper
This shifts all of the logic that was happening in the Playlist script handler into the tab helper. This will allow the tab helper to be used when the `UseProfileWebViewConfiguration` feature flag is enabled
