+++ ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+TabObserver.swift
@@ -417,7 +417,7 @@ extension BrowserViewController {
 
     if tab.profile.prefs.isPlaylistAvailable {
       injectedScripts.append(contentsOf: [
-        PlaylistScriptHandler(tab: tab)
+        PlaylistScriptHandler()
       ])
     }
 
@@ -465,9 +465,6 @@ extension BrowserViewController {
       )
     }
 
-    (tab.browserData?.getContentScript(name: PlaylistScriptHandler.scriptName)
-      as? PlaylistScriptHandler)?
-      .delegate = self
     (tab.browserData?.getContentScript(name: Web3NameServiceScriptHandler.scriptName)
       as? Web3NameServiceScriptHandler)?.delegate = self
   }
